refactor: bootstrap @stll/business-registries umbrella, migrate ARES
Introduce a single workspace package, @stll/business-registries, that hosts
typed clients for national business / commercial registries under per-
jurisdiction subpaths. Move the existing @stll/ares package into
src/ares/, retire packages/ares/, and update the two consumers in
apps/api to import from @stll/business-registries/ares.

No behaviour change. The umbrella exists so the four jurisdictions queued
next (NO Brreg, FI PRH, PL KRS, UK Companies House, FR INSEE Sirene) can
land as folder-add PRs without each spinning up its own package.json /
tsconfig / lint config.

BOE stays as @stll/boe because it is a gazette (legislation + daily
bulletin + BORME), not a business registry.
