+++ apps/api/package.json
@@ -54,8 +54,8 @@
     "@stll/anonymize-chat": "workspace:*",
     "@stll/anonymize-data": "catalog:",
     "@stll/anonymize-wasm": "catalog:",
-    "@stll/ares": "workspace:*",
     "@stll/boe": "workspace:*",
+    "@stll/business-registries": "workspace:*",
     "@stll/case-law": "workspace:*",
     "@stll/docx-core": "workspace:*",
     "@stll/docx-utils": "workspace:*",
