+++ packages/business-registries/src/ares/errors.ts
