+++ packages/business-registries/src/ares/validation.ts
